social_media_api/requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
sqlparse==0.5.4
